reorganize cli
